PMA-127 - applied authorization for post related policies

+++ app/Http/Controllers/PostApprovalBlacklistUnitController.php
@@ -3,12 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\DbModels\PostApprovalBlacklistUnit;
+use App\DbModels\PostWall;
 use App\Http\Requests\PostApprovalBlacklistUnit\IndexRequest;
 use App\Http\Requests\PostApprovalBlacklistUnit\StoreRequest;
 use App\Http\Requests\PostApprovalBlacklistUnit\UpdateRequest;
 use App\Http\Resources\PostApprovalBlacklistUnitResource;
 use App\Http\Resources\PostApprovalBlacklistUnitResourceCollection;
 use App\Repositories\Contracts\PostApprovalBlacklistUnitRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class PostApprovalBlacklistUnitController extends Controller
 {
@@ -31,9 +33,12 @@ class PostApprovalBlacklistUnitController extends Controller
      *
      * @param IndexRequest $request
      * @return PostApprovalBlacklistUnitResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [PostApprovalBlacklistUnit::class, $request->get('propertyId')]);
+
         $postApprovalBlacklistUnits = $this->postApprovalBlacklistUnitRepository->findBy($request->all());
 
         return new PostApprovalBlacklistUnitResourceCollection($postApprovalBlacklistUnits);
@@ -44,9 +49,12 @@ class PostApprovalBlacklistUnitController extends Controller
      *
      * @param  StoreRequest $request
      * @return PostApprovalBlacklistUnitResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [PostApprovalBlacklistUnit::class, $request->get('propertyId')]);
+
         $postApprovalBlacklistUnit = $this->postApprovalBlacklistUnitRepository->save($request->all());
 
         return new PostApprovalBlacklistUnitResource($postApprovalBlacklistUnit);
@@ -57,9 +65,12 @@ class PostApprovalBlacklistUnitController extends Controller
      *
      * @param PostApprovalBlacklistUnit $postApprovalBlacklistUnit
      * @return PostApprovalBlacklistUnitResource
+     * @throws AuthorizationException
      */
     public function show(PostApprovalBlacklistUnit $postApprovalBlacklistUnit)
     {
+        $this->authorize('show', $postApprovalBlacklistUnit);
+
         return new PostApprovalBlacklistUnitResource($postApprovalBlacklistUnit);
     }
 
@@ -69,9 +80,12 @@ class PostApprovalBlacklistUnitController extends Controller
      * @param UpdateRequest $request
      * @param PostApprovalBlacklistUnit $postApprovalBlacklistUnit
      * @return PostApprovalBlacklistUnitResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, PostApprovalBlacklistUnit $postApprovalBlacklistUnit)
     {
+        $this->authorize('update', $postApprovalBlacklistUnit);
+
         $postApprovalBlacklistUnit = $this->postApprovalBlacklistUnitRepository->update($postApprovalBlacklistUnit,$request->all());
 
         return new PostApprovalBlacklistUnitResource($postApprovalBlacklistUnit);
@@ -82,9 +96,12 @@ class PostApprovalBlacklistUnitController extends Controller
      *
      * @param PostApprovalBlacklistUnit $postApprovalBlacklistUnit
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(PostApprovalBlacklistUnit $postApprovalBlacklistUnit)
     {
+        $this->authorize('destroy', $postApprovalBlacklistUnit);
+
         $this->postApprovalBlacklistUnitRepository->delete($postApprovalBlacklistUnit);
 
         return response()->json(null,204);
