+++ app/Http/Controllers/PostCommentController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\PostComment\UpdateRequest;
 use App\Http\Resources\PostCommentResource;
 use App\Http\Resources\PostCommentResourceCollection;
 use App\Repositories\Contracts\PostCommentRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class PostCommentController extends Controller
 {
@@ -31,9 +32,12 @@ class PostCommentController extends Controller
      *
      * @param IndexRequest $request
      * @return PostCommentResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [PostComment::class, $request->get('propertyId')]);
+
         $postComments = $this->postCommentRepository->findBy($request->all());
 
         return new PostCommentResourceCollection($postComments);
@@ -44,9 +48,12 @@ class PostCommentController extends Controller
      *
      * @param StoreRequest $request
      * @return PostCommentResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [PostComment::class, $request->get('propertyId')]);
+
         $postComment = $this->postCommentRepository->save($request->all());
 
         return new PostCommentResource($postComment);
@@ -57,9 +64,12 @@ class PostCommentController extends Controller
      *
      * @param PostComment $postComment
      * @return PostCommentResource
+     * @throws AuthorizationException
      */
     public function show(PostComment $postComment)
     {
+        $this->authorize('show', $postComment);
+
         return new PostCommentResource($postComment);
     }
 
@@ -69,9 +79,12 @@ class PostCommentController extends Controller
      * @param UpdateRequest $request
      * @param PostComment $postComment
      * @return PostCommentResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, PostComment $postComment)
     {
+        $this->authorize('update', $postComment);
+
         $postComment = $this->postCommentRepository->update($postComment, $request->all());
 
         return new PostCommentResource($postComment);
@@ -82,9 +95,12 @@ class PostCommentController extends Controller
      *
      * @param PostComment $postComment
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(PostComment $postComment)
     {
+        $this->authorize('destroy', $postComment);
+
         $this->postCommentRepository->delete($postComment);
 
         return response()->json(null,204);
